+++ app-sample/proguard-rules.pro
