+++ app-sample/src/main/res/layout/activity_dynamic_component.xml
