+++ app-sample/src/main/res/mipmap-xxhdpi/ic_launcher.webp
