+++ app-sample/src/main/res/values-night/themes.xml
