+++ app-sample/src/main/res/values/colors.xml
