+++ app-sample/src/main/res/values/themes.xml
