+++ app-sample/src/main/res/xml/network_security_config.xml
