--- app/src/main/AndroidManifest.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application
-        android:allowBackup="true"
-        android:networkSecurityConfig="@xml/network_security_config"
-        android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
-        android:supportsRtl="true"
-        android:theme="@style/Theme.DynamicView"
-        android:name="com.github.rviannaoliveira.dynamicview.CustomApplication"
-        android:exported="false">
-        <activity
-            android:name="com.github.rviannaoliveira.dynamicview.xml.presentation.CraftDComponentActivity"
-            android:exported="true">
-
-        </activity>
-
-        <activity
-            android:name="com.github.rviannaoliveira.dynamicview.compose.presentation.DynamicComposeActivity"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-    </application>
-
-</manifest>
