--- app/src/main/java/com/github/rviannaoliveira/dynamicview/data/DynamicSampleService.kt
@@ -1,9 +1,0 @@
-package com.github.rviannaoliveira.dynamicview.data
-
-import com.github.codandotv.craftd.androidcore.data.model.base.DataSimplePropertiesResponse
-import retrofit2.http.GET
-
-interface DynamicSampleService {
-    @GET("dynamic")
-    suspend fun getDynamicExample() : DataSimplePropertiesResponse
-}
