+++ settings.gradle
@@ -8,6 +8,6 @@ dependencyResolutionManagement {
     }
 }
 rootProject.name = "DynamicView"
-include ':app'
 include ':dynamic'
 include ':craftd-core'
+include ':app-sample'
